start can test on linux

+++ com/as.virtual/include/Virtio.h
@@ -284,7 +284,7 @@ public:
             vr.avail->ring[free_head] = free_head;
             num_added ++;
             added = true;
-            ASLOG(VRING,"add_buf(%X) vr.avail->idx=%d\n",(int)data,vr.avail->idx);
+            ASLOG(VRING,"add_buf(%X) vr.avail->idx=%d\n",(int)(unsigned long)data,vr.avail->idx);
         }
         else
         {
